+++ walkers/src/tiles.rs
@@ -3,9 +3,8 @@ use std::{collections::HashMap, sync::Arc};
 
 use egui::{pos2, Color32, Context, Mesh, Rect, Vec2};
 use egui_extras::RetainedImage;
-use futures::StreamExt;
-use reqwest::header::USER_AGENT;
 
+use crate::download::download_wrap;
 use crate::io::Runtime;
 use crate::mercator::TileId;
 use crate::providers::{Attribution, TileSource};
@@ -16,7 +15,7 @@ pub struct Tile {
 }
 
 impl Tile {
-    fn from_image_bytes(image: &[u8]) -> Result<Self, String> {
+    pub(crate) fn from_image_bytes(image: &[u8]) -> Result<Self, String> {
         RetainedImage::from_image_bytes("debug_name", image).map(|image| Self {
             image: Arc::new(image),
         })
@@ -110,81 +109,6 @@ impl Tiles {
                 None
             }
         }
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-enum Error {
-    #[error(transparent)]
-    Http(reqwest::Error),
-
-    #[error("error while decoding the image: {0}")]
-    Image(String),
-}
-
-async fn download_single(client: &reqwest::Client, url: &str) -> Result<Tile, Error> {
-    let image = client
-        .get(url)
-        .header(USER_AGENT, "Walkers")
-        .send()
-        .await
-        .map_err(Error::Http)?;
-
-    log::debug!("Downloaded {:?}.", image.status());
-
-    let image = image
-        .error_for_status()
-        .map_err(Error::Http)?
-        .bytes()
-        .await
-        .map_err(Error::Http)?;
-
-    Tile::from_image_bytes(&image).map_err(Error::Image)
-}
-
-async fn download<S>(
-    source: S,
-    mut request_rx: futures::channel::mpsc::Receiver<TileId>,
-    mut tile_tx: futures::channel::mpsc::Sender<(TileId, Tile)>,
-    egui_ctx: Context,
-) -> Result<(), ()>
-where
-    S: TileSource + Send + 'static,
-{
-    // Keep outside the loop to reuse it as much as possible.
-    let client = reqwest::Client::new();
-
-    loop {
-        let request = request_rx.next().await.ok_or(())?;
-        let url = source.tile_url(request);
-
-        log::debug!("Getting {:?} from {}.", request, url);
-
-        match download_single(&client, &url).await {
-            Ok(tile) => {
-                tile_tx.try_send((request, tile)).map_err(|_| ())?;
-                egui_ctx.request_repaint();
-            }
-            Err(e) => {
-                log::warn!("Could not download '{}': {}", &url, e);
-            }
-        }
-    }
-}
-
-async fn download_wrap<S>(
-    source: S,
-    request_rx: futures::channel::mpsc::Receiver<TileId>,
-    tile_tx: futures::channel::mpsc::Sender<(TileId, Tile)>,
-    egui_ctx: Context,
-) where
-    S: TileSource + Send + 'static,
-{
-    if download(source, request_rx, tile_tx, egui_ctx)
-        .await
-        .is_err()
-    {
-        log::error!("Error from IO runtime.");
     }
 }
